Remove avatar and add link to new UI version in MainLayout
Removed avatar and chatbot description from the MainLayout component. Added a section in the same component to inform users about the new UI version of the application and provide them with a link to access this improved version. This gives users access to enhanced functionality while reassuring them of ongoing development support.